@@ -6,19 +6,20 @@
 	"author": "Logan (EthyMoney)",
 	"repository": "https://github.com/YoloSwagDogDiggity/Covid-19",
 	"license": "MIT",
-  "engines": {
-    "npm": ">=8.5.0",
-    "node": ">=18.1.0"
-  },
+  "type": "module",
+	"engines": {
+		"npm": ">=9.1.0",
+		"node": ">=18.12.1"
+	},
 	"dependencies": {
-		"chalk": "^4.1.2",
-		"discord.js": "^13.7.0",
-		"jsdom": "^19.0.0",
-		"node-schedule": "^2.1.0",
-		"shortcountrynames": "^0.9.0",
+		"chalk": "^5.2.0",
+		"discord.js": "^13.12.0",
+		"jsdom": "^21.1.0",
+		"node-schedule": "^2.1.1",
+		"shortcountrynames": "^0.10.0",
 		"topgg-autoposter": "^2.0.1"
 	},
 	"devDependencies": {
-		"eslint": "^8.16.0"
+		"eslint": "^8.33.0"
 	}
 }
